seperated each order into its own component
