+++ src/components/orders/Order.css
@@ -1,0 +1,59 @@
+.order {
+    display:flex;
+    border: 3px var(--accent) solid;
+    border-radius: 15px;
+    padding: 10px;
+    margin: 5px;
+}
+
+.order-details {
+    display:flex;
+    flex-direction: column;
+    flex-basis: 80%;
+    text-decoration: none;
+    color: var(--darkest)
+}
+
+.order-header {
+    display:flex;
+    justify-content: space-evenly;
+    flex-basis:30%;
+}
+
+.order-body {
+    display:flex;
+    flex-direction:column;
+    flex-basis:70%;
+}
+
+.order-table {
+    display:flex;
+    justify-content: center;
+}
+
+.order-employee-taken, .order-employee-delivery {
+    display:flex;
+}
+
+.order-employee-taken-label, .order-employee-delivery-label {
+    flex-basis:60%;
+}
+
+.order-employee-taken-data, .order-employee-delivery-data {
+    flex-basis:40%;
+}
+
+.order-btn-container {
+    display:flex;
+    flex-direction:column;
+    flex-basis:15%;
+    justify-content: space-evenly;
+}
+
+.order-btn-edit {
+    flex-basis:30%;
+}
+
+.order-btn-delete {
+    flex-basis:30%;
+}
