+++ src/components/orders/ViewOrders.jsx
@@ -1,13 +1,12 @@
 import { useEffect, useState } from "react";
-import { deleteOrder, getAllEmployeeOrders, getAllOrders } from "../../services/orderService";
+import { getAllEmployeeOrders, getAllOrders } from "../../services/orderService";
 import "./ViewOrders.css"
-import { Link, useNavigate } from "react-router-dom";
+import { Order } from "./Order";
 
 
 export const ViewOrders = () => {
     const [date, setDate] = useState(new Date().toISOString().slice(0,10))
     const [orders, setOrders] = useState([])
-    const navigate = useNavigate()
     const [employeeOrders, setEmployeeOrders] = useState([])
 
     const getAndSetOrders = () => {
@@ -45,65 +44,7 @@ export const ViewOrders = () => {
             </section>
             <ul className="orders">
                 {orders.map((order) => 
-                    <li className="order" value={order.id} key={order.id}>
-                        <Link to={order.id.toString()} className="order-details" >
-                            <div className="order-header">
-                                <div className="order-number">
-                                    Order #{order.id}
-                                </div>
-                            </div>
-                            <div className="order-body">
-                                <div className="order-employee-taken">
-                                    <div className="order-employee-taken-label">
-                                        Taken By :                                        
-                                    </div>
-                                    <div className="order-employee-taken-data">
-                                        {employeeOrders.find((empOrder) => 
-                                            empOrder.order?.id == order.id && empOrder.tookOrder
-                                        )?.employee?.fullName}
-                                    </div>
-                                </div>
-                                {!order.tableNumber ? (
-                                    <div className="order-employee-delivery">
-                                        <div className="order-employee-delivery-label">
-                                            Assigned to Delivery :
-                                        </div>
-                                        <div className="order-employee-delivery-data">
-                                            {employeeOrders.find((empOrder) =>
-                                                empOrder.order?.id == order.id && !empOrder.tookOrder
-                                            )?.employee?.fullName}
-                                        </div>
-                                    </div>
-                                ) : (
-                                    <div className="order-table">
-                                        Table #{order.tableNumber}
-                                    </div>
-                                )}
-                            </div>
-                        </Link>
-                        <div className="order-btn-container">
-                            <button 
-                                className="order-btn-edit"
-                                value={order.id}
-                                onClick={(event) => {
-                                    navigate(`../orders/edit/${event.target.value}`)
-                                }}
-                            >
-                                Edit
-                            </button>
-                            <button 
-                                className="order-btn-delete"
-                                value={order.id}
-                                onClick={(event) => {
-                                    deleteOrder(event.target.value).then(() => {
-                                        getAndSetOrders()
-                                    })
-                                }}
-                            >
-                                Delete
-                            </button>
-                        </div>
-                    </li>
+                    <Order key={order.id} order={order} employeeOrders={employeeOrders} getAndSetOrders={getAndSetOrders} />
                 )}
             </ul>
         </div>
